refactor: restructure checking the requirements
